Refactored the Owner and pet models into new folder location
