+++ app/src/main/java/com/example/endeavoringorms/persistence/room/entities/one_to_many/Owner.kt
@@ -1,4 +1,4 @@
-package com.example.endeavoringorms.persistence.room.entities
+package com.example.endeavoringorms.persistence.room.entities.one_to_many
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
